+++ tests/test_dataclasses.py
@@ -1,18 +1,11 @@
-import contextlib
-import io
-from typing import Optional
+from dataclasses import dataclass
+from typing import Optional, Any
 
 import pandas as pd
 import pytest
 
 from typeddfs.builders import TypedDfBuilder
-from typeddfs.abs_dfs import AbsDf, TypedDfDataclass
-from typeddfs.utils import FrozeList
-
-
-class _UnhashableUnorderable:
-    def __init__(self, v):
-        self.v = v
+from typeddfs.typed_dfs import TypedDfDataclass
 
 
 class TestDataclasses:
@@ -24,57 +17,88 @@ class TestDataclasses:
         assert dc.get_fields() == []
 
     def test_create_simple(self):
-        t = TypedDfBuilder("T").require("hi", dtype=str).reserve("val", dtype=int).build()
+        t = TypedDfBuilder("T").require("greeting", dtype=str).reserve("bonus", dtype=int).build()
         dc = t.create_dataclass()
         assert dc is not None
         assert issubclass(dc, TypedDfDataclass)
         assert len(dc.get_fields()) == 2
-        assert dc.get_fields()[0].name == "hi"
+        assert dc.get_fields()[0].name == "greeting"
         assert dc.get_fields()[0].type == str
-        assert dc.get_fields()[1].name == "val"
+        assert dc.get_fields()[1].name == "bonus"
         assert dc.get_fields()[1].type == Optional[int]
 
-    def test_create_ordered(self):
-        t = TypedDfBuilder("T").require("hi", dtype=str).reserve("val", dtype=int).build()
+    def test_create_any(self):
+        t = TypedDfBuilder("T").require("greeting").reserve("bonus").build()
         dc = t.create_dataclass()
+        assert dc is not None
         assert issubclass(dc, TypedDfDataclass)
-        assert hasattr(dc, "__lt__") and dc.__lt__ is not None
-
-    def test_create_unordered(self):
-        t = TypedDfBuilder("T").require("hi", dtype=list).reserve("val", dtype=int).build()
-        dc = t.create_dataclass()
-        assert issubclass(dc, TypedDfDataclass)
-        # assert dc.get_df_type() is t  # TODO
-        # noinspection PyArgumentList
-        x = dc(hi="cat", val=1)
-        assert isinstance(x, TypedDfDataclass)
-        # noinspection PyUnresolvedReferences
-        assert x.hi == "cat"
-        # noinspection PyUnresolvedReferences
-        assert x.val == 1
-        assert hasattr(dc, "__lt__") and dc.__lt__ is not None
+        assert len(dc.get_fields()) == 2
+        assert dc.get_fields()[0].name == "greeting"
+        assert dc.get_fields()[0].type == Any
+        assert dc.get_fields()[1].name == "bonus"
+        assert dc.get_fields()[1].type == Optional[Any]
 
     def test_to_instances(self):
-        t = TypedDfBuilder("T").require("hi", dtype=str).reserve("val", dtype=int).build()
-        df: t = t.of([pd.Series(dict(hi="dog", val=111))])
+        t = TypedDfBuilder("T").require("animal", dtype=str).reserve("age", dtype=int).build()
+        df: t = t.of(
+            [
+                pd.Series(dict(animal="goldfish", age=2)),
+                pd.Series(dict(animal="goldfish", age=1)),
+                pd.Series(dict(animal="gazelle", age=8)),
+                pd.Series(dict(animal="pineapple", age=114)),
+                pd.Series(dict(animal="anteater", age=11)),
+            ]
+        )
         dc = t.create_dataclass()
-        inst = df.to_dataclass_instances()
-        assert len(inst) == 1
-        assert inst[0].hi == "dog"
-        assert inst[0].val == 111
-        assert inst[0] == dc(hi="dog", val=111)
+        instances = df.to_dataclass_instances()
+        assert instances == [
+            dc("goldfish", 2),
+            dc("goldfish", 1),
+            dc("gazelle", 8),
+            dc("pineapple", 114),
+            dc("anteater", 11),
+        ]
+        assert list(sorted(instances)) == [
+            dc("anteater", 11),
+            dc("gazelle", 8),
+            dc("goldfish", 1),
+            dc("goldfish", 2),
+            dc("pineapple", 114),
+        ]
 
-    def test_to_instances_frozen(self):
-        t = TypedDfBuilder("T").require("hi", dtype=str).require("vals").build()
-        df: t = t.of([pd.Series(dict(hi="dog", vals=[1, 2]))])
-        dc = t.create_dataclass()
-        inst = df.to_dataclass_instances()
-        assert len(inst) == 1
-        assert inst[0].hi == "dog"
-        assert inst[0].vals == FrozeList([1, 2])
-        assert inst[0].vals == [1, 2]
-        assert inst[0] == dc(hi="dog", vals=FrozeList([1, 2]))
-        assert inst[0] == dc(hi="dog", vals=[1, 2])  # technically works
+    def test_to_instances_empty(self):
+        t = TypedDfBuilder("T").reserve("animal", dtype=str).build()
+        df: t = t.of([])
+        instances = df.to_dataclass_instances()
+        assert instances == []
+
+    def test_read_instances(self):
+        @dataclass(
+            frozen=True,
+        )
+        class Dc:
+            animal: str
+            val: Optional[int]
+
+        t = TypedDfBuilder("T").require("animal", dtype=str).reserve("age", dtype=int).build()
+        df = t.from_dataclass_instances([Dc("cat", 1), Dc("kitten", 2)])
+        assert len(df) == 2
+        assert df.values.tolist() == [["cat", 1], ["kitten", 2]]
+
+    def test_read_instances_empty(self):
+        t = TypedDfBuilder("T").require("animal", dtype=str).build()
+        df = t.from_dataclass_instances([])
+        assert len(df) == 0
+
+    def test_read_instances_empty_fields(self):
+        @dataclass(frozen=True)
+        class Dc:
+            pass
+
+        t = TypedDfBuilder("T").reserve("animal", dtype=str).build()
+        df = t.from_dataclass_instances([Dc()])
+        assert len(df) == 1
+        assert "animal" not in df.columns
 
 
 if __name__ == "__main__":